refactor: remove useless source
