+++ src/pgs_defs.h
@@ -1,7 +1,7 @@
 #ifndef _PGS_DEFS
 #define _PGS_DEFS
 
-#include <string.h> // memset
+#include <string.h>
 
 #define BUFSIZE_16K 16 * 1024
 #define memzero(buf, n) (void)memset(buf, 0, n)
