+++ test/CMakeLists.txt
@@ -4,7 +4,6 @@ include_directories (${CMAKE_SOURCE_DIR}/src)
 add_executable (pgs_util_test pgs_util_test.c 
                        ${CMAKE_SOURCE_DIR}/3rd-party/sha3.c
                        ${CMAKE_SOURCE_DIR}/3rd-party/hash_32a.c
-                       ${CMAKE_SOURCE_DIR}/src/pgs_util.c
                        ${CMAKE_SOURCE_DIR}/src/pgs_crypto.c)
 target_link_libraries (pgs_util_test ${OPENSSL_LIBRARIES})
